feat: Vercel AI SDK integration
