+++ .gitignore
@@ -16,3 +16,5 @@ __pycache__/
 .pytest_cache/
 
 .env
+
+dist/
