+++ examples/ai-sdk/.env.example
@@ -1,0 +1,4 @@
+POSTHOG_API_TOKEN=phx_your_token_here
+POSTHOG_API_BASE_URL=https://us.posthog.com
+INKEEP_API_KEY=your_inkeep_key_here
+OPENAI_API_KEY=sk-your_openai_key_here
