+++ examples/ai-sdk/package.json
@@ -1,0 +1,22 @@
+{
+	"name": "posthog-ai-sdk-example",
+	"version": "1.0.0",
+	"description": "Example using PostHog tools with the AI SDK",
+	"type": "module",
+	"scripts": {
+		"dev": "tsx src/index.ts",
+		"start": "node dist/index.js"
+	},
+	"dependencies": {
+		"@posthog/agents": "file:../../typescript",
+		"ai": "^5.0.18",
+		"@ai-sdk/openai": "^2.0.16",
+		"zod": "^3.24.4",
+		"dotenv": "^16.4.7"
+	},
+	"devDependencies": {
+		"@types/node": "^22.15.34",
+		"tsx": "^4.20.3",
+		"typescript": "^5.8.3"
+	}
+}
