+++ examples/ai-sdk/src/index.ts
@@ -1,0 +1,61 @@
+import { openai } from "@ai-sdk/openai";
+import { PostHogAgentToolkit } from "@posthog/agents/integrations/ai-sdk";
+import { generateText, stepCountIs } from "ai";
+import "dotenv/config";
+
+async function analyzeProductUsage() {
+	console.log("🚀 PostHog AI Agent - Product Usage Analysis\n");
+
+	const agentToolkit = new PostHogAgentToolkit({
+		posthogApiToken: process.env.POSTHOG_API_TOKEN!,
+		posthogApiBaseUrl: process.env.POSTHOG_API_BASE_URL || "https://us.posthog.com",
+		inkeepApiKey: process.env.INKEEP_API_KEY,
+	});
+
+	const result = await generateText({
+		model: openai("gpt-5-mini"),
+		tools: agentToolkit.getTools(),
+		stopWhen: stepCountIs(5),
+		system: `You are a data analyst. Your task is to do a deep dive into what's happening in our product.`,
+		prompt: `Please analyze our product usage:
+        
+        1. Get all available insights (limit 100) and pick the 5 most relevant ones
+        2. For each insight, query its data
+        3. Summarize the key findings in a brief report
+        
+        Keep your response focused and data-driven.`,
+	});
+
+	console.log("📊 Analysis Complete!\n");
+	console.log("=".repeat(50));
+	console.log(result.text);
+	console.log("=".repeat(50));
+
+	// Show tool usage summary
+	const toolCalls = result.steps.flatMap((step) => step.toolCalls ?? []);
+	if (toolCalls.length > 0) {
+		console.log("\n🔧 Tools Used:");
+		const toolUsage = toolCalls.reduce(
+			(acc, call) => {
+				acc[call.toolName] = (acc[call.toolName] || 0) + 1;
+				return acc;
+			},
+			{} as Record<string, number>,
+		);
+
+		for (const [tool, count] of Object.entries(toolUsage)) {
+			console.log(`  • ${tool}: ${count} call${count > 1 ? "s" : ""}`);
+		}
+	}
+}
+
+async function main() {
+	try {
+		await analyzeProductUsage();
+	} catch (error) {
+		console.error("Error:", error);
+		process.exit(1);
+	}
+}
+
+main().catch(console.error);
