+++ typescript/package.json
@@ -1,0 +1,51 @@
+{
+	"name": "@posthog/agents",
+	"version": "0.1.0",
+	"description": "PostHog MCP tools and integrations for AI agents",
+	"main": "dist/index.js",
+	"module": "dist/index.mjs",
+	"types": "dist/index.d.ts",
+	"exports": {
+		".": {
+			"types": "./dist/index.d.ts",
+			"import": "./dist/index.mjs",
+			"require": "./dist/index.js"
+		},
+		"./tools": {
+			"types": "./dist/tools.d.ts",
+			"import": "./dist/tools.mjs",
+			"require": "./dist/tools.js"
+		},
+		"./integrations/ai-sdk": {
+			"types": "./dist/ai-sdk.d.ts",
+			"import": "./dist/ai-sdk.mjs",
+			"require": "./dist/ai-sdk.js"
+		}
+	},
+	"scripts": {
+		"build": "tsup",
+		"dev": "tsup --watch",
+		"format": "biome format --write",
+		"lint": "biome lint --fix",
+		"typecheck": "tsc --noEmit"
+	},
+	"keywords": ["posthog", "mcp", "ai", "agents", "analytics", "feature-flags"],
+	"author": "PostHog Inc.",
+	"license": "MIT",
+	"peerDependencies": {
+		"ai": "^5.0.0",
+		"zod": "^3.24.0"
+	},
+	"dependencies": {
+		"posthog-node": "^4.18.0"
+	},
+	"devDependencies": {
+		"@biomejs/biome": "^1.9.4",
+		"@types/node": "^22.15.34",
+		"ai": "^5.0.18",
+		"tsup": "^8.5.0",
+		"typescript": "^5.8.3",
+		"zod": "^3.24.4"
+	},
+	"files": ["dist", "README.md"]
+}
