+++ typescript/src/tools/dashboards/addInsight.ts
@@ -9,7 +9,7 @@ type Params = z.infer<typeof schema>;
 
 export const addInsightHandler = async (context: Context, params: Params) => {
 	const { data } = params;
-	const projectId = await context.getProjectId();
+	const projectId = await context.stateManager.getProjectId();
 
 	const insightResult = await context.api
 		.insights({ projectId })
