+++ typescript/src/tools/toolDefinitions.ts
@@ -2,6 +2,8 @@ import toolDefinitionsJson from "../../../schema/tool-definitions.json";
 
 export interface ToolDefinition {
 	description: string;
+	category?: string;
+	summary?: string;
 }
 
 export type ToolDefinitions = Record<string, ToolDefinition>;
@@ -12,8 +14,10 @@ export default toolDefinitions;
 
 export function getToolDefinition(toolName: string): ToolDefinition {
 	const definition = toolDefinitions[toolName];
+
 	if (!definition) {
 		throw new Error(`Tool definition not found for: ${toolName}`);
 	}
+
 	return definition;
 }
