+++ wrangler.jsonc
@@ -5,7 +5,7 @@
 {
 	"$schema": "node_modules/wrangler/config-schema.json",
 	"name": "mcp1",
-	"main": "typescript/src/index.ts",
+	"main": "typescript/src/integrations/mcp/index.ts",
 	"compatibility_date": "2025-03-10",
 	"compatibility_flags": ["nodejs_compat"],
 	"migrations": [
